fix(T11898): make pi single-envelope/exit-trap tests in-process (drop fragile tsx subprocess)
The two pi smoke tests spawned `.ts` fixtures via a `tsx` loader. Resolving that
loader is fragile under CI's hoisted-pnpm layout: tsx is not on PATH, and tsx 4.x
does not export the `dist/cli.mjs` subpath earlier attempts reached, so the
exports-map resolution still threw "Unable to resolve the tsx loader" on both the
ubuntu and macOS shards. The child never started — status null, zero envelope
lines.

Both behaviors are fully observable in-process, so the subprocess (and tsx) is
gone:

- single-envelope: capture process.stdout.write/process.stderr.write around a
  REAL PiAgentAdapter.run() + the REAL createEnvelope writer; assert exactly ONE
  LAFS-envelope line hit stdout and the streaming noise went to stderr.
- daemon-survives: drive the REAL S1 containment (installDaemonExitGuard +
  wrapPiCall) with a fn that calls process.exit(1); assert it THROWS the typed
  PiContainmentError (E_PI_PROCESS_EXIT_TRAPPED) and that this process is still
  alive (reaching the assertion = survived). No child, no subprocess exit-code.

Deletes the two now-orphaned fixtures. Assertions are unchanged in WHAT they
prove: exactly-one-envelope-on-stdout + a contained, survived forced exit. All
66 pi tests pass in a cgroup-wrapped run and under a tsx-stripped PATH.
